feat: add Dockerfile and optimize for containers

+++ src/lib/server/database.ts
@@ -2,14 +2,19 @@ import 'dotenv/config';
 import { PrismaBetterSqlite3 } from '@prisma/adapter-better-sqlite3';
 import { PrismaClient } from '../prisma/client';
 import { env } from '$env/dynamic/private';
+import { building } from '$app/env';
 
 const connectionString = env.DATABASE_URL ?? 'file:./dev.db';
 
 const adapter = new PrismaBetterSqlite3({ url: connectionString });
-const client = new PrismaClient({ adapter });
+export const client = new PrismaClient({ adapter });
 
-await client.floatingOrder.deleteMany({
-	where: {}
+if (!building) {
+	await client.floatingOrder.deleteMany({
+		where: {}
+	});
+}
+
+process.on('sveltekit:shutdown', async (reason) => {
+	await client.$disconnect();
 });
-
-export { client };
